Merge official Stockfish commits from the Move class change up to the last commit
before the global variable refactor. Decisions taken in this step:
- Move becomes a class as upstream, but with 32-bit storage and Fairy's encoding
  (7-bit squares, move type, promotion/drop/gating piece type, gating square or
  in-hand piece type). Fairy's free helper functions (from_sq(m), type_of(m),
  gating_square(m), make_drop(), ...) are kept as wrappers so variant code and
  bindings keep compiling; MOVE_NONE/MOVE_NULL alias Move::none()/Move::null().
- Value becomes an int alias as upstream; all Fairy piece value constants are
  now constexpr Values. Score and the enum operator macros for other types stay.
- BAD_QUIET move picker stage and correction history tweaks adopted.
- Dual NNUE (big/small net) is skipped: upstream removed it again before SF 19
  and Fairy keeps its own single-network architecture.
- TT entries keep the 32-bit move.

+++ src/tt.cpp
@@ -35,8 +35,8 @@ TranspositionTable TT;  // Our global transposition table
 void TTEntry::save(Key k, Value v, bool pv, Bound b, Depth d, Move m, Value ev) {
 
     // Preserve any existing move for the same position
-    if (m || (uint16_t) k != key16)
-        move32 = (uint32_t) m;
+    if (m || uint16_t(k) != key16)
+        move32 = m;
 
     // Overwrite less valuable entries (cheapest checks first)
     if (b == BOUND_EXACT || (uint16_t) k != key16 || d - DEPTH_OFFSET + 2 * pv > depth8 - 4)
